# Reduce container drawable overhead from comparer creation