Just a minor fix (on paper saves 24 bytes or so). Was more interested in getting some kind of benchmark coverage in place.

before:
|              Method |     Mean |   Error |  StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------- |---------:|--------:|--------:|------:|------:|------:|----------:|
|   TestEmptyGameLoop | 994.7 us | 2.21 us | 2.07 us |     - |     - |     - |      41 B |
|  TestAllocContainer | 993.5 us | 2.17 us | 2.03 us |     - |     - |     - |    2001 B |
| TestLoadedContainer | 990.8 us | 2.12 us | 1.98 us |     - |     - |     - |    2428 B |
|     TestAllocSprite | 993.1 us | 2.63 us | 2.46 us |     - |     - |     - |    1601 B |
|    TestLoadedSprite | 993.1 us | 1.53 us | 1.43 us |     - |     - |     - |    1853 B |

after:
|              Method |     Mean |   Error |  StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------- |---------:|--------:|--------:|------:|------:|------:|----------:|
|   TestEmptyGameLoop | 994.5 us | 2.52 us | 2.36 us |     - |     - |     - |      41 B |
|  TestAllocContainer | 993.4 us | 2.03 us | 1.90 us |     - |     - |     - |    1996 B |
| TestLoadedContainer | 993.0 us | 1.57 us | 1.47 us |     - |     - |     - |    2404 B |
|     TestAllocSprite | 992.3 us | 7.51 us | 7.02 us |     - |     - |     - |    1601 B |
|    TestLoadedSprite | 993.8 us | 1.66 us | 1.56 us |     - |     - |     - |    1853 B |